backend/requirements.txt
asgiref==3.7.2
awscli==1.29.30
beautifulsoup4==4.12.2
bleach==6.0.0
botocore==1.31.30
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.4
contourpy==1.0.7
crypto==1.4.1
cycler==0.11.0
Django==4.2.3
django-cors-headers==4.2.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
docutils==0.16
drf-yasg==1.21.7
fonttools==4.39.4
idna==3.4
inflection==0.5.1
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.3
matplotlib==3.7.1
mpmath==1.3.0
Naked==0.1.32
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
pbkdf2==1.3
Pillow==10.0.0
pyasn1==0.5.0
pycrypto==2.6.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.28.2
rsa==4.7.2
s3transfer==0.6.2
shellescape==3.8.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
